requirement.txt
Flask
pdfplumber
python-dotenv
faiss-cpu
google-generativeai
langchain-google-genai
langchain
